split out bill and testimony search components
