+++ components/search/index.ts
@@ -1,2 +1,3 @@
-export * from "./Search"
 export * from "./service"
+export * from "./BillSearch"
+export * from "./TestimonySearch"
